implement support for appending signatures with resign
The main use cases for multiple signatures on one message are:

* Simplify CA certificate roll-over by signing messages with keys in
  both the old and new hierarchies during a transition period.
  Recipients would have either the old or new CA installed (or
  temporarily even both). In either case, we'd like to accept any
  message which has at least one signature that can be verified.

* Support user-implemented verification policies which require
  signatures from multiple specific signers.

* Support user-implemented verification policies which require a minimum
  number of signatures from different signers under under a trusted CA.

As a first step, we add a --append option for 'rauc resign'.

We call CMS_add1_signer with CMS_REUSE_DIGEST so that we don't need
access to the bundle payload for plain bundles. As we always use the
same digest algorithm (SHA256), this is not a problem for us. If no
matching digest is found, OpenSSL will raise an error.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>