config_file: prepare to split out target checks from load_config
Currently, load_config checks that the config has all required field for
use on the target. We want to support -C options even without a full
config, so we need to split config parsing from these checks.

To make it easier to review, first create a check_config_target() helper
which is called from load_config.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>