test/test_bundle: drop -progress from --mksquashfs-args
Since d52f951 [1], which is part of squashfs-tools >= 4.7, a check
exists that not both options -progress and -no-progress are set.

Since RAUC sets -no-progress by default, the check for
--mksquashfs-args, which additionally sets -progress, starts to cause an
error with recent mksquashfs versions:

| ERROR: Only one of -no-progress and -progress can be specified.  Both causes a conflict.

Fix this by just not setting -progress in the test case, since the
actual option we set doesn't matter.

[1] https://github.com/plougher/squashfs-tools/commit/d52f9513f061cf8c76206c639e47d8afcceb8bb8

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>