test/update_handler: extend get_update_handler tests
Improve test coverage to ensure that the availability of all currently
possible update pair combinations is tested.
Add test cases for the following combinations:
- img.caibx to nor, vfat, and jffs2 filesystems
- img to jffs2 filesystem

The 'img_to_nand_handler' is currently marked as explicitly
unsupported, but it is still implicitly supported via the fallback
handler '{"*.img.caibx", "*", img_to_raw_handler}'. To ensure that future
refactoring does not introduce behavioral changes a test case for this
combination is also added.

Signed-off-by: Lars Schmidt <l.schmidt@pengutronix.de>