src/main: limit scope of --keyring argument
The --keyring argument is supported by most, but not all arguments.
And its actual interpretation depends on the subcommand.

E.g. for 'rauc install' with the service enabled, it is simply a no-op,
which can be quite irritating for users and is not ideal for
security-related context.

As a preparation, we hide the --keyring option from the global option
context (G_OPTION_FLAG_HIDDEN) and prevent auto-aliasing for the
duplicate local options (G_OPTION_FLAG_NOALIAS).
This aligns with what we did for other options like --cert or --key in
the past, too.

Note that this does not change the parsing behavior, yet, to stay
compatible with option ordering supported so far (and used in tests).

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>