add 'bundle name' to event log
Add the bundle (file) name, that is passed into rauc to update the
system with, to the event log message.

This way a the original filename - and not just the bundle hash - is
available to find the corresponding bundle that e.g. lead to an error
that has is being debugged.

Signed-off-by: Johannes Schneider <johannes.schneider@leica-geosystems.com>